# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mini Wallet System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /transfers/initiate:
    post:
      operationId: initiateTransfer
      tags: [transfers]
      summary: Initiate a wallet-to-wallet transfer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferBody"
      responses:
        "202":
          description: Transfer initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate request (idempotency)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResult"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transfers/{transactionId}/status:
    get:
      operationId: getTransferStatus
      tags: [transfers]
      summary: Get transfer status by transaction ID
      security:
        - BearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transfer status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications for current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security:
        - BearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - BearerAuth: []
      responses:
        "200":
          description: All marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /simulate/db-failure:
    post:
      operationId: simulateDbFailure
      tags: [simulation]
      summary: Simulate a database failure
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResult"
  /simulate/duplicate-payment:
    post:
      operationId: simulateDuplicatePayment
      tags: [simulation]
      summary: Simulate a duplicate payment (idempotency test)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DuplicatePaymentBody"
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResult"
  /simulate/notification-failure:
    post:
      operationId: simulateNotificationFailure
      tags: [simulation]
      summary: Simulate a notification service failure
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResult"
  /simulate/kafka-delay:
    post:
      operationId: simulateKafkaDelay
      tags: [simulation]
      summary: Simulate Kafka processing delay
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KafkaDelayBody"
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResult"
  /users/search:
    get:
      operationId: searchUsers
      tags: [auth]
      summary: Search users by email for transfers
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSearchResponse"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    RegisterBody:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        walletId:
          type: string
      required: [id, name, email, role, createdAt]
...
    TransferBody:
      type: object
      properties:
        receiverEmail:
          type: string
          format: email
        amount:
          type: number
          minimum: 1
        note:
          type: string
        idempotencyKey:
          type: string
      required: [receiverEmail, amount]
    TransactionResult:
      type: object
      properties:
        transactionId:
          type: string
        status:
          type: string
          enum: [INITIATED, PROCESSING, SUCCESS, FAILED, PENDING, REVERSED]
        amount:
          type: number
        message:
          type: string
        idempotencyKey:
          type: string
        isDuplicate:
          type: boolean
      required: [transactionId, status, amount, message]
    Transaction:
      type: object
      properties:
        id:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        senderEmail:
          type: string
        receiverId:
          type: string
        receiverName:
          type: string
        receiverEmail:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [INITIATED, PROCESSING, SUCCESS, FAILED, PENDING, REVERSED]
        type:
          type: string
          enum: [TRANSFER, ADD_MONEY, WITHDRAW, REFUND]
        note:
          type: string
        idempotencyKey:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, amount, status, type, createdAt, updatedAt]
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [EMAIL, SMS, IN_APP]
        status:
          type: string
          enum: [SENT, DELIVERED, FAILED, READ]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, message, type, status, createdAt]
    NotificationListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: integer
      required: [notifications, unreadCount]
...
    SimulationResult:
      type: object
      properties:
        scenario:
          type: string
        description:
          type: string
        outcome:
          type: string
        steps:
          type: array
          items:
            type: string
        compensationApplied:
          type: boolean
      required: [scenario, description, outcome, steps]
    DuplicatePaymentBody:
      type: object
      properties:
        idempotencyKey:
          type: string
        amount:
          type: number
      required: [idempotencyKey, amount]
    KafkaDelayBody:
      type: object
      properties:
        delayMs:
          type: integer
          minimum: 100
          maximum: 10000
      required: [delayMs]
    UserSearchResult:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
      required: [id, name, email]
    UserSearchResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserSearchResult"
      required: [users]